pyproject.toml
[project]
name = "kokomoor-platform"
version = "0.1.0"
description = "Personal agentic pipeline ecosystem — shared infrastructure for job search, ML showcase, and future automation pipelines."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Samuel Kokomoor", email = "kokomoor@mit.edu" }]

dependencies = [
    # LLM
    "anthropic>=0.39.0",
    # Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # Browser automation
    "playwright>=1.48.0",
    # Data layer
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    # Validation & config
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    # Observability
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    # Notifications
    "aiosmtplib>=3.0.0",
    # Document generation
    "python-docx>=1.1.0",
    "PyYAML>=6.0",
    # Utilities
    "httpx>=0.27.0",
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    # Scraper pipeline
    "mmh3>=4.0.0",
    "aioimaplib>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",
    "respx>=0.21.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "pipelines*"]

# ---------- Ruff ----------
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["core", "pipelines", "scripts"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["core", "pipelines"]

# ---------- Mypy ----------
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["playwright.*", "langgraph.*", "langchain_core.*", "docx.*", "bs4.*"]
ignore_missing_imports = true

# ---------- Pytest ----------
[tool.pytest.ini_options]
testpaths = ["core", "pipelines"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
